+++ R/FuzzyHits-class.R
@@ -6,7 +6,7 @@
 #' @aliases membership,FuzzyHits-method
 #' @importFrom methods slot
 setMethod("membership", "FuzzyHits", function(x) {
-    slot(x, "membership")
+    elementMetadata(x)[["membership"]]
 })
 
 #' @param value An object of a class specified in the S4 method signature or as outlined in 'Slots'.
@@ -15,15 +15,14 @@ setMethod("membership", "FuzzyHits", function(x) {
 #' @aliases membership<-,FuzzyHits-method
 #' @importFrom methods validObject
 setMethod("membership<-", "FuzzyHits", function(x, value) {
-    x@membership <- value
-    validObject(x)
+    elementMetadata(x)[["membership"]] <- value
     x
 })
 
 # show() ----
 
 setMethod("show", "FuzzyHits", function(object) {
-    elementMetadata(object) <- DataFrame(membership=membership(object))
+    elementMetadata(object) <- elementMetadata(object)[, "membership", drop=FALSE]
     S4Vectors:::showHits(object, margin = "  ", print.classinfo = TRUE, print.nnode = TRUE)
 })
 
@@ -31,6 +30,11 @@ setMethod("show", "FuzzyHits", function(object) {
 
 #' @importFrom methods new
 setAs("Hits", "FuzzyHits", function(from) {
-    new("FuzzyHits", from, membership=rep(1, length(from)))
+    if (! "membership" %in% colnames(elementMetadata(from))) {
+        elementMetadata(from)[["membership"]] <- rep(1, length(from))
+    }
+    from <- subset(from, !is.na(membership))
+    to <- new("FuzzyHits", from)
+    to
 })