+++ R/FuzzySet-class.R
@@ -3,18 +3,9 @@
 #' @param x An object that inherits from `FuzzySets`.
 #'
 #' @rdname FuzzySets-class
-#' @aliases relations,FuzzySets-method
-#' @importFrom S4Vectors DataFrame
-setMethod("relations", "FuzzySets", function(x) {
-    out <- relations(as(x, "BaseSets"))
-    out$membership <- membership(x)
-    out
-})
-
-#' @rdname FuzzySets-class
 #' @aliases membership,FuzzySets-method
 setMethod("membership", "FuzzySets", function(x) {
-    membership(x@relations)
+    as.numeric(membership(x@relations))
 })
 
 #' @param value An object of a class specified in the S4 method signature or as outlined in 'Slots'.
@@ -33,24 +24,19 @@ setMethod("membership<-", "FuzzySets", function(x, value) {
 setMethod("subset", "FuzzySets", function(x, ...) {
     .local <- function (x, subset, select, drop=FALSE, ...) {
         # Only difference with the BaseSets parent method
-        table <- cbind(as(x, "data.frame"), membership=membership(x@relations))
+        table <- as(x, "data.frame")
         i <- eval(substitute(subset), table)
 
         keep.element <- unique(id(elementData(x))[from(x@relations)[i]])
         keep.set <- unique(id(setData(x))[to(x@relations)[i]])
 
-        relations <- DataFrame(table[i, c("element", "set"), drop=FALSE])
+        relations <- DataFrame(table[i, , drop=FALSE])
         elementData <- elementData(x)[which(id(elementData(x)) %in% keep.element)]
         setData <- setData(x)[which(id(setData(x)) %in% keep.set)]
-        membership <- membership(x@relations)[i]
+        membership <- membership(x)[i]
 
-        fs <- BaseSets(relations, elementData, setData)
-        # Upgrade Hits to FuzzyHits
-        relations <- new("FuzzyHits", fs@relations, membership=membership)
-        fs@relations <- relations
-        fs <- new("FuzzySets", fs)
-        validObject(fs)
-        fs
+        out <- FuzzySets(relations, elementData, setData)
+        out
     }
     .local(x, ...)
 })
@@ -58,21 +44,22 @@ setMethod("subset", "FuzzySets", function(x, ...) {
 # show() ----
 
 setMethod("show", "FuzzySets", function(object) {
-    # Combine elementData, setData, and relations into a single DataFrame
-    element <- elementData(object)[from(object@relations)]
-    elementData <- elementMetadata(element)
-    elementMetadata(element) <- NULL # avoid metadata columns
-    set <- setData(object)[to(object@relations)]
-    setData <- elementMetadata(set)
-    elementMetadata(set) <- NULL # avoid metadata columns
-    x <- DataFrame(
-        element=element,
-        set=set
-    )
-    x[["membership"]] <- membership(object@relations)
-    x[["elementData"]] <- elementData
-    x[["setData"]] <- setData
-
+    # element <- elementData(object)[from(object@relations)]
+    # set <- setData(object)[to(object@relations)]
+    # setData <- elementMetadata(set)
+    # elementMetadata(set) <- NULL # avoid metadata columns
+    # Combine element and set identifiers into a DataFrame
+    x <- as(object, "DataFrame")
+    # Retain only the core columns
+    x <- x[, c("element", "set"), drop=FALSE]
+    # Show the rest of the relation metadata
+    x[["relationData"]] <- elementMetadata(object@relations)
+    # Show the element metadata
+    elementData <- elementData(object)[from(object@relations)]
+    x[["elementData"]] <-  elementMetadata(elementData)
+    # Show the set metadata
+    setData <- setData(object)[to(object@relations)]
+    x[["setData"]] <- elementMetadata(setData)
     .showSetAsTable(class(object), x)
 })
 
@@ -102,9 +89,9 @@ as.list.FuzzySets <- function(x, ...) {
 #'
 #' as.matrix(fs, fun.aggregate=min)
 #'
-as.matrix.FuzzySets <- function(x, ..., fill=0) {
+as.matrix.FuzzySets <- function(x, ..., fill=NA_real_) {
     out <- as(x, "data.frame")
-    out[["value"]] <- membership(x@relations)
+    out[["value"]] <- membership(x)
     out <- acast(out, element~set, value.var="value", fill=fill, ...)
     out
 }
@@ -118,12 +105,9 @@ setAs("FuzzySets", "matrix", function(from) {
 #' @importFrom S4Vectors DataFrame
 setAs("matrix", "FuzzySets", function(from) {
     storage.mode(from) <- "double"
-    x <- melt(from, varnames=c("element", "set"), as.is=TRUE)
-    x <- x[!is.na(x$value), , drop=FALSE]
-    relations <- x[, c("element", "set"), drop=FALSE]
+    relations <- melt(from, varnames=c("element", "set"), value.name = "membership", as.is=TRUE)
     relations <- DataFrame(relations)
-    membership <- x$value
-    FuzzySets(relations, membership=membership)
+    FuzzySets(relations)
 })
 
 #' @aliases as.FuzzySets.matrix as.FuzzySets
@@ -134,12 +118,9 @@ as.FuzzySets.matrix <- function(x, ...) {
 
 #' @importFrom methods new
 setAs("BaseSets", "FuzzySets", function(from) {
-    fuzzyhits <- as(from@relations, "FuzzyHits")
-    membership(fuzzyhits) <- rep(1, nRelations(from))
-    from@relations <- fuzzyhits
-    x <- new("FuzzySets", from)
-    validObject(x)
-    x
+    from@relations <- as(from@relations, "FuzzyHits")
+    to <- new("FuzzySets", from)
+    to
 })
 
 #' @aliases as.FuzzySets.BaseSets as.FuzzySets