+++ tests/testthat/test-FuzzyHits.R
@@ -11,32 +11,42 @@ names(membership) <- paste0("x", seq_along(membership))
 
 test_that("FuzzyHits constructor produces valid objects", {
 
-    expect_message(FuzzyHits(from, to, 7, 15, membership), "Setting names(membership) to NULL", fixed=TRUE)
+    expect_message(
+        FuzzyHits(from, to, membership, 7, 15),
+        "Setting names(membership) to NULL",
+        fixed=TRUE)
 
-    out <- FuzzyHits(from, to, 7, 15, membership)
+    out <- FuzzyHits(from, to, membership, 7, 15)
 
     expect_s4_class(out, "FuzzyHits")
+    expect_length(out, length(from))
 })
 
 test_that("FuzzyHits validity method identifies issues", {
 
     # Invalid membership length
     expect_error(
-        FuzzyHits(from=c(1, 2), to=c(1, 2), 7, 15, membership=c(0.4)),
-        "'membership(x)' is not parallel to 'x'",
+        FuzzyHits(from=c(1, 2), to=c(1, 2), membership=c(0.4), 7, 15),
+        "mcols(x)' is not parallel to 'x'",
         fixed=TRUE
     )
 
     # membership function out of range [0,1]
     expect_error(
-        FuzzyHits(from=c(1), to=c(1), 7, 15, membership=c(-1)),
+        FuzzyHits(from=c(1), to=c(1), membership=c(-1), 7, 15),
         "membership function must be in the interval [0,1]",
         fixed=TRUE
     )
 
     expect_error(
-        FuzzyHits(from=c(1), to=c(1), 7, 15, membership=c(2)),
+        FuzzyHits(from=c(1), to=c(1), membership=c(2), 7, 15),
         "membership function must be in the interval [0,1]",
+        fixed=TRUE
+    )
+
+    expect_error(
+        FuzzyHits(from=c(1), to=c(1), membership=c(NA), 7, 15),
+        "membership function must be numeric",
         fixed=TRUE
     )
 
@@ -46,9 +56,10 @@ test_that("FuzzyHits validity method identifies issues", {
 
 test_that("show(FuzzySets) works", {
 
-    fs <- FuzzyHits(from, to, 7, 15, membership)
+    fs <- FuzzyHits(from, to, membership, 7, 15)
 
     out <- show(fs)
     expect_null(out)
 
 })
+